fix: fix incorrect use of MediaType.get()
By adding the correct argument for MediaType a bunch more bugs
showed themselfs in the system and there for had to be fixed
as well.

The code section "when(client.newCall(bookSearch).execute())"
would cause NullPointerException cause the return value for
"client.newCall(bookSearch)" had not been defined. SO, setupCall
was created to solve this issue.